Implemented ability to display IMDB data on Details page.
Added GetImdbData method to MediaController. Modified Details action to check if MediaObject is derived from DatabaseSource 1 (IMDB) and passed an OMDbTitle to EditMediaObjectViewModel, which was modified to accept. Modified OMDbTitle to collect new data for display. Modified Details view to display new data if it is a movie, series, or episode.